requirements.txt
Flask==2.0.1
pyaudio==0.2.11
numpy
python-dotenv==0.19.0
torch==2.1.0
whisperplus  # This might need to be adjusted based on the actual package name and version
wave